Add style for the `/talktoyourself` marker
When using the command `/talktoyourself on`, an indicator appears above the chat
input textbox.  This was showing as white on white, and was impossible to read.
I've made some guesses as to reasonable styling to fix this, giving it the same
background color as the area around the text input box, and changing the text to
yellow (because it should be fairly obvious when it's turned on).